Added functionality to write calendars to 'calendar.ics' file in root directory
