+++ calendar.ics
@@ -1,0 +1,25 @@
+0201201T120000-0500
+SUMMARY:Math Class
+DESCRIPTION:Algebra II, Room 203
+LOCATION:Walter Reed Middle School
+END:VEVENT
+BEGIN:VEVENT
+UID:20201201T130000-0500-02345-10001
+DTSTAMP:20201201T130000-0500
+DTSTART:20201201T130000-0500
+DTEND:20201201T140000-0500
+SUMMARY:Science Class
+DESCRIPTION:Biology, Room 101
+LOCATION:Walter Reed Middle School
+END:VEVENT
+BEGIN:VEVENT
+UID:20201201T150000-0500-02345-10002
+DTSTAMP:20201201T150000-0500
+DTSTART:20201201T150000-0500
+DTEND:20201201T160000-0500
+SUMMARY:History Class
+DESCRIPTION:World History, Room 305
+LOCATION:Walter Reed Middle School
+END:VEVENT
+END:VCALENDAR
+```
