+++ package.json
@@ -12,7 +12,8 @@
     "@google-cloud/vertexai": "^0.3.0",
     "next": "14.1.0",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
